Add helper class for bag load on/off
